[bugfix] always do Ember.defineProperty
Certain libraries rely on Ember hooks when defining properties,
most notably EmberData. By always running Ember.defineProperty,
we can ensure that computed properties are always defined exactly
the same way as Ember does it by default.

Note that this still means that normal class props will not go
through this process. I think this is ultimately preferable, we
will likely want to drop this custom behavior as we can, but
this exception makes sense for the moment.